Add a bunch of useful methods to RTree/Node.
In particular, this commit introduces searchIterator, which returns
an iterator over the search results. This method could be used to
implement most of the other search-based methods.

We need to do some efficiency testing to see which methods are
the fastest. The goal is to minimize the number of traversal methods
in Node while still preserving performance for the important ones.